Use withText() functionality in the search-header app
This functionality was deveoped within the team to avoid prop drilling
inside apps, which we encountered around the project. For more info
about this functionality - refer to:
https://github.com/danskernesdigitalebibliotek/dpl-react/pull/27
In a nutshell, we wrap the app in a HOC and all the props are saved into
the redux state, which can be accessed from subcomponents.
The search header app and autosuggest, though were built before withText
was made. That's why we now introduce it in the app retrospectively.
