+++ src/apps/search-header/search-header.entry.tsx
@@ -1,22 +1,23 @@
 import * as React from "react";
-import SearchHeader, { SearchHeaderProps } from "./search-header";
+import { withText } from "../../core/utils/text";
+import SearchHeader from "./search-header";
+
+export interface SearchHeaderProps {
+  searchHeaderUrlText?: string;
+  altText?: string;
+  inputPlaceholderText?: string;
+  stringSuggestionAuthorText?: string;
+  stringSuggestionWorkText?: string;
+  stringSuggestionTopicText?: string;
+}
 
 const SearchHeaderEntry: React.FC<SearchHeaderProps> = ({
-  searchHeaderUrl,
-  altText,
-  inputPlaceholderText,
-  stringSuggestionAuthorText,
-  stringSuggestionWorkText,
-  stringSuggestionTopicText
-}) => (
-  <SearchHeader
-    searchHeaderUrl={searchHeaderUrl}
-    altText={altText}
-    inputPlaceholderText={inputPlaceholderText}
-    stringSuggestionAuthorText={stringSuggestionAuthorText}
-    stringSuggestionWorkText={stringSuggestionWorkText}
-    stringSuggestionTopicText={stringSuggestionTopicText}
-  />
-);
+  searchHeaderUrlText = "https://bibliotek.dk/search",
+  altText = "search icon",
+  inputPlaceholderText = "Search here",
+  stringSuggestionAuthorText = "author",
+  stringSuggestionWorkText = "work",
+  stringSuggestionTopicText = "topic"
+}) => <SearchHeader />;
 
-export default SearchHeaderEntry;
+export default withText(SearchHeaderEntry);
