+++ src/components/search-bar/search-bar.tsx
@@ -1,19 +1,14 @@
 import * as React from "react";
 import searchIcon from "@danskernesdigitalebibliotek/dpl-design-system/build/icons/basic/icon-search.svg";
 import { UseComboboxPropGetters } from "downshift";
+import { useText } from "../../core/utils/text";
 
 export interface SearchBarProps {
-  searchHeaderUrl?: string;
-  altText?: string;
-  inputPlaceholderText?: string;
   getInputProps: UseComboboxPropGetters<unknown>["getInputProps"];
 }
 
-const SearchBar: React.FC<SearchBarProps> = ({
-  altText = "search icon",
-  inputPlaceholderText = "Search here",
-  getInputProps
-}) => {
+const SearchBar: React.FC<SearchBarProps> = ({ getInputProps }) => {
+  const t = useText();
   return (
     <>
       {/* eslint-disable react/jsx-props-no-spreading */}
@@ -23,14 +18,14 @@ const SearchBar: React.FC<SearchBarProps> = ({
         className="header__menu-search-input text-body-medium-regular"
         type="text"
         autoComplete="off"
-        placeholder={inputPlaceholderText}
+        placeholder={t("inputPlaceholderText")}
         {...getInputProps()}
       />
       {/* eslint-enable react/jsx-props-no-spreading */}
       <input
         type="image"
         src={searchIcon}
-        alt={altText}
+        alt={t("altText")}
         className="header__menu-search-icon"
       />
     </>
